# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cars Project Documentation",
        "description": "Cars OpenApi description",
        "contact": {
            "email": "boxtab@gmail.com"
        },
        "license": {
            "name": "Car License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://car.local/api/v1",
            "description": "Car API Server"
        }
    ],
    "paths": {
        "/admin/car/list": {
            "get": {
                "tags": [
                    "car"
                ],
                "summary": "list",
                "operationId": "741346986e2806f349ec1f4ab0dfbbd4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "brand": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 4,
                                                    "Brand": "Audi"
                                                },
                                                {
                                                    "id": 5,
                                                    "Brand": "Chevrolet"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/car": {
            "post": {
                "tags": [
                    "car"
                ],
                "summary": "create",
                "operationId": "db1b714f7ac0e97759c8987210584969",
                "requestBody": {
                    "description": "Pass car credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand"
                                ],
                                "properties": {
                                    "brand": {
                                        "type": "string",
                                        "example": "Mercedes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "My intenal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/admin/car/{id}": {
            "get": {
                "tags": [
                    "car"
                ],
                "summary": "read",
                "operationId": "647b640fce63757ca28198d4119e5c36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": {
                                                "id": 3,
                                                "brand": "Ford"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/user/list": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "list",
                "operationId": "76d630aadfcdcdca81f305945b49b075",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "test-name",
                                                    "email": "test1@gmail.com"
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "test2",
                                                    "email": "test2@gamil.com"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "404": {
                "description": "fail"
            }
        }
    },
    "tags": [
        {
            "name": "Projects Car",
            "description": "API Endpoints of cars"
        }
    ]
}